# fs: allow pre/background-fetch to be called before layer verification