--- a/fs/fs.go
+++ b/fs/fs.go
@@ -177,6 +177,13 @@ func (fs *filesystem) Mount(ctx context.Context, mountpoint string, labels map[s
 		return fmt.Errorf("source must be passed")
 	}
 
+	defaultPrefetchSize := fs.prefetchSize
+	if psStr, ok := labels[config.TargetPrefetchSizeLabel]; ok {
+		if ps, err := strconv.ParseInt(psStr, 10, 64); err == nil {
+			defaultPrefetchSize = ps
+		}
+	}
+
 	// Resolve the target layer
 	var (
 		resultChan = make(chan layer.Layer)
@@ -188,10 +195,10 @@ func (fs *filesystem) Mount(ctx context.Context, mountpoint string, labels map[s
 			l, err := fs.resolver.Resolve(ctx, s.Hosts, s.Name, s.Target)
 			if err == nil {
 				resultChan <- l
+				fs.prefetch(ctx, l, defaultPrefetchSize, start)
 				return
 			}
-			rErr = errors.Wrapf(rErr, "failed to resolve layer %q from %q: %v",
-				s.Target.Digest, s.Name, err)
+			rErr = errors.Wrapf(rErr, "failed to resolve layer %q from %q: %v", s.Target.Digest, s.Name, err)
 		}
 		errChan <- rErr
 	}()
@@ -202,12 +209,17 @@ func (fs *filesystem) Mount(ctx context.Context, mountpoint string, labels map[s
 		desc := desc
 		go func() {
 			// Avoids to get canceled by client.
-			ctx := log.WithLogger(context.Background(),
-				log.G(ctx).WithField("mountpoint", mountpoint))
-			err := fs.resolver.Cache(ctx, preResolve.Hosts, preResolve.Name, desc)
+			ctx := log.WithLogger(context.Background(), log.G(ctx).WithField("mountpoint", mountpoint))
+			l, err := fs.resolver.Resolve(ctx, preResolve.Hosts, preResolve.Name, desc)
 			if err != nil {
 				log.G(ctx).WithError(err).Debug("failed to pre-resolve")
+				return
 			}
+			fs.prefetch(ctx, l, defaultPrefetchSize, start)
+
+			// Release this layer because this isn't target and we don't use it anymore here.
+			// However, this will remain on the resolver cache until eviction.
+			l.Done()
 		}()
 	}
 
@@ -270,41 +282,6 @@ func (fs *filesystem) Mount(ctx context.Context, mountpoint string, labels map[s
 	fs.layer[mountpoint] = l
 	fs.layerMu.Unlock()
 	fs.metricsController.Add(mountpoint, l)
-
-	// Prefetch this layer. We prefetch several layers in parallel. The first
-	// Check() for this layer waits for the prefetch completion.
-	if !fs.noprefetch {
-		prefetchSize := fs.prefetchSize
-		if psStr, ok := labels[config.TargetPrefetchSizeLabel]; ok {
-			if ps, err := strconv.ParseInt(psStr, 10, 64); err == nil {
-				prefetchSize = ps
-			}
-		}
-		go func() {
-			fs.backgroundTaskManager.DoPrioritizedTask()
-			defer fs.backgroundTaskManager.DonePrioritizedTask()
-			if err := l.Prefetch(prefetchSize); err != nil {
-				log.G(ctx).WithError(err).Warnf("failed to prefetch layer=%v", digest)
-				return
-			}
-			log.G(ctx).Debug("completed to prefetch")
-		}()
-	}
-
-	// Fetch whole layer aggressively in background. We use background
-	// reader for this so prioritized tasks(Mount, Check, etc...) can
-	// interrupt the reading. This can avoid disturbing prioritized tasks
-	// about NW traffic.
-	if !fs.noBackgroundFetch {
-		go func() {
-			if err := l.BackgroundFetch(); err != nil {
-				log.G(ctx).WithError(err).Warnf("failed to fetch whole layer=%v", digest)
-				return
-			}
-			commonmetrics.LogLatencyForLastOnDemandFetch(ctx, digest, start, l.Info().ReadTime) // write log record for the latency between mount start and last on demand fetch
-			log.G(ctx).Debug("completed to fetch all layer data in background")
-		}()
-	}
 
 	// mount the node to the specified mountpoint
 	// TODO: bind mount the state directory as a read-only fs on snapshotter's side
@@ -420,6 +397,23 @@ func (fs *filesystem) Unmount(ctx context.Context, mountpoint string) error {
 	// goroutine using channel, etc.
 	// See also: https://www.kernel.org/doc/html/latest/filesystems/fuse.html#aborting-a-filesystem-connection
 	return syscall.Unmount(mountpoint, syscall.MNT_FORCE)
+}
+
+func (fs *filesystem) prefetch(ctx context.Context, l layer.Layer, defaultPrefetchSize int64, start time.Time) {
+	// Prefetch a layer. The first Check() for this layer waits for the prefetch completion.
+	if !fs.noprefetch {
+		go l.Prefetch(defaultPrefetchSize)
+	}
+
+	// Fetch whole layer aggressively in background.
+	if !fs.noBackgroundFetch {
+		go func() {
+			if err := l.BackgroundFetch(); err == nil {
+				// write log record for the latency between mount start and last on demand fetch
+				commonmetrics.LogLatencyForLastOnDemandFetch(ctx, l.Info().Digest, start, l.Info().ReadTime)
+			}
+		}()
+	}
 }
 
 // neighboringLayers returns layer descriptors except the `target` layer in the specified manifest.