@rdpsin 2021-09-20
Thank you very much! I tested out the patch and prefetch pull times reduced by upto 3x for some images. 😄 

I'm curious about why there is still latency during prefetch and if we can do something to bring it down even further. Thanks again!