@ktock 2021-09-22
> Thank you very much! I tested out the patch and prefetch pull times reduced by upto 3x for some images. smile

Thank you for testing this patch!

> I'm curious about why there is still latency during prefetch and if we can do something to bring it down even further. 

Prefetch starts after [`layer.Resolve()`](https://github.com/containerd/stargz-snapshotter/blob/3fe9ff9a78cc3ae3d214184ba977b162305b5322/fs/fs.go#L213). Though `layer.Resolve()` of layers start simultaneously, the time to take for `layer.Resolve()` seems different among layers and this seems to cause the difference of start latency of prefetch.